Clean harmonization implementation
