+++ multiqc_config.yaml
@@ -9,9 +9,6 @@ streamline_count:
 metricsinroi:
   iqr_multiplier: 3 # Default: samples outside Q1 - 3*IQR to Q3 + 3*IQR are marked as fail
 
-metricsinroi_harmonized:
-  iqr_multiplier: 3 # Default: samples outside Q1 - 3*IQR to Q3 + 3*IQR are marked as fail
-
 # Framewise Displacement module configuration default thresholds
 framewise_displacement:
   warn_threshold: 0.8 # Default: samples with max FD < 0.8mm get "pass" status
@@ -52,13 +49,12 @@ report_section_order:
 
 # Order that modules should appear in the report
 module_order:
-  - harmonization
-  - metricsinroi_harmonized
   - framewise_displacement
   - coverage
   - streamline_count
   - tractometry
   - metricsinroi
+  - harmonization
   - cortical
   - subcortical
